+++ shared/index.js
@@ -1,5 +1,6 @@
 export const TRANSPORT_UNINITIALIZED = 'Transport uninitialized';
 export const UI = '127.0.0.1:8180';
+export const DAPPS = '127.0.0.1:8080';
 export const ACCOUNTS_REQUEST = 'parity.background.accounts';
 
 /**
