+++ web3/index.js
@@ -16,9 +16,43 @@ if (window.location.protocol === 'chrome-extension:') {
   const script = document.createElement('script');
   script.src = chrome.extension.getURL('web3/inpage.js');
   document.documentElement.insertBefore(script, document.documentElement.childNodes[0]);
-  const port = chrome.runtime.connect({ name: 'web3' });
+
+  const initPort = () => {
+    const port = chrome.runtime.connect({ name: 'web3' });
+    if (!port) {
+      return;
+    }
+
+    port.onMessage.addListener((msg) => {
+      const { id, err, payload } = msg;
+
+      // Inject iframe only if the page is using Web3
+      if (!payload || payload.id !== ACCOUNTS_REQUEST) {
+        if (!err) {
+          injectIframe();
+        } else {
+          // remove iframe
+          removeIframe(err);
+        }
+      }
+
+      window.postMessage({
+        type: 'parity.web3.response',
+        id,
+        err,
+        payload
+      }, '*');
+    });
+
+    port.onDisconnect.addListener(() => {
+      port.isDisconnected = true;
+    });
+
+    return port;
+  };
 
   // process requests
+  let port = initPort();
   window.addEventListener('message', (ev) => {
     if (ev.source !== window) {
       return;
@@ -31,6 +65,13 @@ if (window.location.protocol === 'chrome-extension:') {
     const { type } = ev.data;
 
     if (type === 'parity.web3.request') {
+      if (!port || port.isDisconnected) {
+        // try to reconnect
+        port = initPort();
+      }
+
+      // add origin information
+      ev.data.origin = window.location.origin;
       port.postMessage(ev.data);
       return;
     }
@@ -42,27 +83,6 @@ if (window.location.protocol === 'chrome-extension:') {
         backgroundSeed: ev.data.backgroundSeed
       });
     }
-  });
-
-  port.onMessage.addListener((msg) => {
-    const { id, err, payload } = msg;
-
-    // Inject iframe only if the page is using Web3
-    if (!payload || payload.id !== ACCOUNTS_REQUEST) {
-      if (!err) {
-        injectIframe();
-      } else {
-        // remove iframe
-        removeIframe(err);
-      }
-    }
-
-    window.postMessage({
-      type: 'parity.web3.response',
-      id,
-      err,
-      payload
-    }, '*');
   });
 }
